+++ src/utils/ua-utils.ts
@@ -2,6 +2,10 @@ export function isFirefox(): boolean {
   return navigator.userAgent.indexOf('Firefox/') !== -1;
 }
 
+export function isFenix(): boolean {
+  return isFirefox() && navigator.userAgent.indexOf('Android') !== -1;
+}
+
 export function isChromium(): boolean {
   return (
     navigator.userAgent.indexOf('Chrome/') !== -1 ||