+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Stop showing the popup for number-only matches in the name dictionary
   (e.g. 64)
   [#811](https://github.com/birchill/10ten-ja-reader/issues/811)).
+- Fix formatting of options page on Firefox for Android (Nightly).
 
 ## 1.4.8 (2021-10-21)